Add files via upload
This is the first code. This also takes into account turning off the coils after moving to a position to reduce energy consumption. 
Using PWM for each PIN would be more elegant since we could avoid de-energizing completely the coil-phase but simply reducing the current. This would guarantee better cogging trque while still reducing consumption... This is a future next step for this library.